adjusting file directory

+++ hc_sr04/tof_noLib/tof_noLib.ino
@@ -8,7 +8,7 @@
 #define echoPin 3      //* UNO input
 #define pulseTimeout  100000  //* 100 milli-sec
 
-double duration;
+long duration;
 long dist_cm;  //* distance in centimeter
 
 //* ---------- ---------- ----------
